% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasum.R
\name{DataSumm}
\alias{DataSumm}
\title{Summarize an Entire Data Frame}
\usage{
DataSumm(data)
}
\arguments{
\item{data}{A data frame (tibble is also ok).}
}
\value{
A data frame, one row per input column.
}
\description{
Applies \code{Datum()} to each column of a data frame and binds the results.
}
\examples{
DataSumm(iris)
}
