% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasum.R
\name{DataSumm}
\alias{DataSumm}
\title{Summarize an Entire Data Frame}
\usage{
DataSumm(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A data frame with summary statistics for each column.
}
\description{
This function summarizes each column of a data frame by calculating various statistics.
}
\examples{
DataSumm(iris)
}
