% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasum.R
\name{Datum}
\alias{Datum}
\title{Summarize a Single Vector}
\usage{
Datum(data)
}
\arguments{
\item{data}{A numeric, character, or factor vector.}
}
\value{
A data frame with summary statistics.
}
\description{
This function summarizes a single vector by calculating various statistics.
}
\examples{
Datum(rnorm(100))
}
