% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasum.R
\name{getmode}
\alias{getmode}
\title{Get Mode of a Numeric Vector}
\usage{
getmode(data)
}
\arguments{
\item{data}{A numeric vector.}
}
\value{
The mode of the numeric vector.
}
\description{
This function calculates the mode of a numeric vector.
}
\examples{
getmode(c(1, 2, 2, 3, 4))
}
