% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasum.R
\name{shapiro_normality_test}
\alias{shapiro_normality_test}
\title{Perform Normality Test}
\usage{
shapiro_normality_test(data)
}
\arguments{
\item{data}{A numeric vector.}
}
\value{
A character string indicating whether the data is "Normal" or "Not Normal".
}
\description{
This function performs the Shapiro-Wilk test if the sample size is between 3 and 5000. Otherwise, it performs the Anderson-Darling test.
}
\examples{
shapiro_normality_test(rnorm(100))
}
